==> Jupyter_Notebooks/1.2_metars_extraction.ipynb <==
# EXTRACCIÓN DE PARTES METAR

## En este Notebook se va a desarrollar el proceso de estracción de todos los partes METAR del Aeropuerto Adolfo Suárez Madrid-Barajas en el periodo comprendido entre el 01-11-2017 y el 31-10-2023.

Un parte o código METAR es el estándar internacional en aviación del formato del código utilizado para emitir periódicamente informes de las observaciones meteorológicas en los aeródromos y aeropuertos.

Se trata de un reporte breve en forma de código alfanumérico que aporta información meteorológica detallada de un momento determinado en un aeropuerto concreto. Básicamente, es una sucesión de letras y números que se emiten periódicamente por los aeropuertos y aeródromos. 

Estos partes se emiten periódicamente, generalmente cada 30 minutos, salvo circunstancias excepcionales en las que se pueden emitir partes adicionales. 

Aunque pueden ser dificíl de descifrar para el usuario general, contienen información de gran relevancia para la actividad aérea relativa a la temperatura, precipitaciones, visibilidad en pista, viento y ráfagas entre otro. Por suerte, se pueden encontrar webs que traducen esta información a un lenguaje más amable y que además mantienen registros de los partes en el tiempo.

## Para ello se va a utilizar la técnica de extracción de datos "webscrapping" sobre la web [tutiempo](https://www.tutiempo.net/registros/lemd) que mantiene registros de estos partes desde hace vários años.


### Las librerías que se van a emplear principalmente para el proceso son:
- **selenium** 
- **joblib**
- **pandas** 


# %pip install joblib

from joblib import Parallel, delayed
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import Select   # seleccion de un dropdown
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import TimeoutException
import time

from selenium.webdriver.support import expected_conditions as EC
import time
import pandas as pd
import multiprocessing as mp
import warnings
warnings.filterwarnings('ignore')

from joblib._parallel_backends import LokyBackend
import asyncio

from tqdm.notebook import tqdm
from datetime import datetime, timedelta

## Se habilitan algunas opciones de interés para el driver

#driver configuration
opciones=Options()

opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.headless=False    # si True, no aperece la ventana (headless=no visible)
opciones.add_argument('--start-maximized')         # comienza maximizado
opciones.add_argument('--incognito')

# Despues de hacer un estudio sobre la estructura de la web para plantear el proceso de extracción, se observa que en cada página se encuentran exclusivamente los partes de un día en concreto( de media unos 48).

## Además, se comprueba que al pasar de un día a otro el *link* de la web varía siguiendo un patrón claro como el siguiente:

### - ...tutiempo.net/records/lemd/{<span style="color:red">dia</span>}-{<span style="color:red">mes</span>}-{<span style="color:red">año</span>}.html

### Este formate de *links* es de especial ayuda, ya que nos va a permitir paralelizar el proceso con **joblib**, reduciendo considerablemente el tiempo de extracción.

## Se generan todos los links para el rango de fechas deseado

# Función para formatear el día eliminando el cero delante de los números del 1 al 9
def formatear_dia(dia):
    return str(dia) if dia >= 10 else f"{dia}"

# Definir la fecha de inicio y fin
fecha_inicio = datetime(2017, 11, 1)
fecha_fin = datetime(2022, 12, 31)

# Lista para almacenar los enlaces generados
enlaces = []

# Bucle para generar los enlaces para cada día en el rango especificado
while fecha_inicio <= fecha_fin:
    # Formatear la fecha según el formato deseado
    formato_fecha = f"{formatear_dia(fecha_inicio.day)}-{fecha_inicio.strftime('%B')}-{fecha_inicio.year}".lower()
    
    # Crear el enlace y añadirlo a la lista
    enlace = f"https://en.tutiempo.net/records/lemd/{formato_fecha}.html"
    enlaces.append(enlace)
    
    # Avanzar al siguiente día
    fecha_inicio += timedelta(days=1)


len(enlaces)

# La siguiente función <span style="color:red">extraer</span>, engloba el proceso de extraccíon de los datos necesarios para cada una de las *urls*.

## Los comentaríos de la función explican paso por paso la lógica del proceso.
## A modo de resumen, el driver entra en la página, *clicka* aceptar en los popups de cookies, extrae los datos necesarios almacenadolos en una tabla y finálmente devuelve un DataFrame con los registro METAR de ese día.

dias = []

def extraer(url):
    table = []  # Inicializa table como una lista vacía
    columns = []  # Inicializa columns como una lista vacía
    
    try:
        # Inicia el driver en la URL indicada
        driver = webdriver.Chrome()
        driver.get(url)
        wait = WebDriverWait(driver, 60)
        
        # Esperar a que aparezca el botón de aceptar normas
        normas_button_xpath = '/html/body/div[18]/div[2]/div[1]/div[2]/div[2]/button[1]'
        wait.until(EC.element_to_be_clickable((By.XPATH, normas_button_xpath)))
        normas_button = driver.find_element(By.XPATH, normas_button_xpath)
        normas_button.click()
        print("Aceptado normas")

        # Esperar a que aparezca el botón de aceptar cookies
        cookies_button_xpath = '//*[@id="DivAceptarCookies"]/div/a[2]'
        wait.until(EC.element_to_be_clickable((By.XPATH, cookies_button_xpath)))
        cookies_button = driver.find_element(By.XPATH, cookies_button_xpath)
        cookies_button.click()
        print("Aceptadas cookies")
        
        # Esperar a que cargue la página (usando un elemento en la tabla como referencia)
        wait.until(EC.presence_of_element_located((By.XPATH, '//table//tbody//tr[3]')))
        print("Página cargada")

        day = driver.find_elements(By.XPATH, '//table//tbody//tr')[1].text

        time.sleep(5)
        table = [
            row.text.split('\n')[0:3] + row.text.replace(' km/h', '').split('\n')[-1].split(' ', 2)
            for row in driver.find_elements(By.XPATH, '//table//tbody//tr')[3::2]
        ]  # Copia los registros
        print("Registros extraídos")
        
        columns = ["Day", "Hour", "Condition", "Temperature", "Wind", "Relative_hum", "Pressure"]  # Añade las columnas

        for i in table:
            i.insert(0, day)  # Inserta los registros en la tabla
            
        return pd.DataFrame(table, columns=columns)  # Devuelve un DataFrame con los datos de ese día
        
    except TimeoutException as te:  # Captura TimeoutException
        print(f"Excepción de tiempo de espera: {te}")
        print(f"No se pudo completar la extracción en la URL: {url}")
        return pd.DataFrame(table, columns=columns) 
    except Exception as e:  # Captura otras excepciones
        print(f"Error: {e}")
        print(f"Error en la URL: {url}")
        return pd.DataFrame(table, columns=columns) 
    finally:
        driver.quit()

## El siguiente código inicia el proceso de extracción usando la paralelización con 6 núcleos, es decir, realiza el proceso en *6 urls* simultáneamente.

 Durante el proceso se comprueba que la web en ocasiones tiene errores con los headers de las tablas que generan problemas en las tablas generadas, por lo que se decide hacer la recolección de año en año para mantener un mejor control sobre los errores e ir solucionandolos.

Ademas, en cada año se pierden(no se terminan de abrir correctamente en torno al 5% de las urls) por lo que se implementa un escrapeo iterativo en el que se anotan los días sin registros y se vuelve a intentar abrir el enlace, hasta obtener los partes del 100% de los días.

%%time
paralelo = Parallel(n_jobs=6, verbose=True)

lst_df = paralelo(delayed(extraer)(url) for url in tqdm(enlaces2))

## Obtenidos los DataFrame para cada unos de los días del mes, se concatenan los DataFrame y se hace una breve exploración sobre el resultado para observar si se han obtenido todos los registros necesarios.

pd.set_option('display.max_rows', None)
metar_2018 = pd.concat(lst_df)
print(len(metar_2018))

metar_2019 = pd.concat(lst_df)
print(len(metar_2019))

metar_2020 = pd.concat(lst_df)
print(len(metar_2020))

metar_2021 = pd.concat(lst_df)
print(len(metar_2021))

metar_2022 = pd.concat(lst_df)
print(len(metar_2022))

lista2020 = list(metar_2020.Day.unique())

lista2021 = list(metar_2021.Day.unique())

lista2022 = list(metar_2022.Day.unique())

## Finalmente se exportan a un archivo *.csv* cada uno de los DataFrames mensuales

metar_2018.to_csv("../data/metars/metars_2018_347.csv", index=False)

metar_2019.to_csv("../data/metars/metars_2019_347.csv", index=False)

metar_2020.to_csv("../data/metars/metars_2020_361.csv", index=False)

metar_2021.to_csv("../data/metars/metars_2021_359.csv", index=False)

metar_2022.to_csv("../data/metars/metars_2022_359.csv", index=False)

metar_2017 = pd.read_csv("../data/metars/metars_2017.csv")

metar_2018 = pd.read_csv("../data/metars/metars_2018_347.csv")

metar_2019 = pd.read_csv("../data/metars/metars_2019_347.csv")

# Concat

metars_18_22 = pd.concat([metar_2017,metar_2018,metar_2019,metar_2020,metar_2021,metar_2022])

metars_18_22=metars_18_22.drop_duplicates()

metars_18_22 = metars_18_22[metars_18_22.Pressure.notna()]

metars_18_22 = metars_18_22.reset_index(drop = True)

len(list(metars_18_22.Day.unique()))

# Convertir la lista existente a objetos datetime
dias_existente = list(metars_18_22.Day.unique())
dias_existente = [datetime.strptime(dia, '%A %d %B %Y') for dia in dias_existente]

# Rango de fechas deseado
fecha_inicio = datetime(2017, 11, 1)
fecha_fin = datetime(2022, 12, 31)

# Generar lista completa de días entre las fechas
dias_completo = [fecha_inicio + timedelta(days=d) for d in range((fecha_fin - fecha_inicio).days + 1)]

# Filtrar los días que no están en la lista existente
dias_faltantes = [dia.strftime('%A %d %B %Y') for dia in dias_completo if dia not in dias_existente]
len(dias_faltantes)

# Convertir la lista de días al formato deseado
dias_formato_fecha = [datetime.strptime(dia, '%A %d %B %Y').strftime('%d-%B-%Y').lower() for dia in dias_faltantes]

# Eliminar el 0 delante de los días del 1 al 9
dias_formato_fecha = [fecha[1:] if fecha.startswith('0') else fecha for fecha in dias_formato_fecha]

# Generar enlaces
enlaces2 = [f"https://en.tutiempo.net/records/lemd/{formato_fecha}.html" for formato_fecha in dias_formato_fecha]
enlaces2

metar_res2 = pd.concat(lst_df)
print(len(metar_res2))

len(metar_res2.Day.value_counts())

metar_res2 = metar_res2.drop_duplicates()

metar_res2 = metar_res2[metar_res1.Pressure.notna()]

metar_res2 = metar_res2.reset_index(drop = True)

metars_18_22 = pd.concat([metars_18_22,metar_res2])

# 2022

metar_2022 = metar_2022.drop_duplicates()

metar_2022 = metar_2022[metar_2022.Pressure.notna()]

metar_2022 = metar_2022.reset_index(drop = True)

# metar_2022.Day.value_counts()

metar_2022.Day.loc[4057:4100] = "Sunday 27 March 2022"

metar_2022.drop([4101,4102], inplace=True)

# metar_2022[metar_2022.Day == "Sunday 27 March 2022"]

# 2021

metar_2021 = metar_2021.drop_duplicates()

metar_2021 = metar_2021[metar_2021.Pressure.notna()]

metar_2021 = metar_2021.reset_index(drop = True)

# metar_2021.Day.value_counts()

metar_2021.drop([4145,4146], inplace=True)

metar_2021.Day.loc[4101:4144] = "Sunday 28 March 2021"

# metar_2021[metar_2021.Day == "Sunday 28 March 2021"]

# 2020

metar_2020 = metar_2020.drop_duplicates()

metar_2020 = metar_2020[metar_2020.Pressure.notna()]

metar_2020 = metar_2020.reset_index(drop = True)

# metar_2020.Day.value_counts()

# metar_2020[metar_2020.Day == "Sunday 29 March 2020"]

metar_2020.drop([4266,4267], inplace=True)

metar_2020.Day.loc[4222:4265] = "Sunday 29 March 2020"

# 2019

metar_2019 = metar_2019.drop_duplicates()

metar_2019 = metar_2019[metar_2019.Pressure.notna()]

metar_2019 = metar_2019.reset_index(drop = True)

metar_2019.Day.loc[4173:4218] = "Monday 1 April 2019"

# metar_2019[metar_2019.Day == "Monday 1 April 2019"]

# metar_2019.Day.value_counts()

# 2018

metar_2018 = metar_2018.drop_duplicates()

metar_2018 = metar_2018.reset_index(drop = True)

metar_2018.Day.loc[3947:3990] = "Sunday 25 March 2018"

metar_2018.drop([3991,3992], inplace=True)

# metar_2018.Day.value_counts()

metar_2018 = metar_2018[metar_2018.Pressure.notna()]

metar_2019.info(memory_usage = "deep")

# Final

metars_18_22.shape

metars_18_22.to_csv("../data/metars/metars_2017_2022.csv", index=False)
