==> Jupyter_Notebooks/3.1_concat_flights.ipynb <==
# En este Notebook se van a concatenar los datos obtenidos en Notebooks anteriores con los del proyecto ETL mencionado en el README para obtener la serie completa.

import numpy as np
import pylab as plt   
import seaborn as sns
import pandas as pd
import warnings
warnings.filterwarnings('ignore')
import sys
sys.path.append('../src')
from funflights import *

fl1 = pd.read_csv("../data/flights/flights2017_2019.csv", encoding = "latin1")
fl2 = pd.read_csv("../data/flights/flights2020_2022.csv", encoding = "latin1")
fl3 = pd.read_csv("../data/flights/flights.csv", encoding = "latin1")

fl1.shape

fl2.shape

fl3.shape

fl = pd.concat([fl1,fl2,fl3])

fl.shape

fl.head()

columnas_excepto_una = fl.columns.difference(['flight_id'])  
fl = fl.drop_duplicates(subset=columnas_excepto_una, keep='first')
fl.shape

fl['flight_id'] = range(1, len(fl) + 1)
fl = fl.reset_index(drop = True)
fl.tail()

pd.set_option('display.max_rows', None)


fl.cod_flight_ICAO[fl.cod_flight_ICAO.isna()] = "-"

pd.set_option('display.max_columns', None)
fl.head()

fl.to_csv("../data/flights/flights_total.csv", index=False)
