==> Jupyter_Notebooks/4.1_cross_tables.ipynb <==
# CRUCE DE LAS TABLAS OBTENIDAS Y LIMPIADAS

## En este Notebook se va a desarrollar el proceso para identificar cada registro de vuelo con el parte METAR anterior a la hora programada de salida de cada vuelo.

Para ello, debido al gran número de registros que se van a "cruzar" entre los diferentes DataFrames, se va a utilizar la librería <span style="color:red">**DASK**</span> por su mejor eficiencia y velocidad respecto a Pandas.

# %pip install "dask[complete]"
# %pip install dask_ml

import time
import numpy as np
import pandas as pd
import dask
import dask.dataframe as dd
import warnings
warnings.filterwarnings('ignore')

## Se cargan en DataFrame de *Dask* los 2 datasets previamente limpiados

fl_dask = dd.read_csv("../data/flights/flights_total.csv",dtype={"arr_mins_of_delay": "object", "dep_mins_of_delay": "object"})
mt_dask = dd.read_csv("../data/metars/metars_2017_2023.csv")


fl_dask.head()

mt_dask.head()

## Se homogeniza el nombre de las columnas por las que se van a cruzar los registros, se ponen en un formato *datetime* común y se ordenan.

fl_dask = fl_dask.rename(columns={'departure_date_time': 'Date_time'})

mt_dask['Date_time'] = dd.to_datetime(mt_dask['Date_time'])
fl_dask['Date_time'] = dd.to_datetime(fl_dask['Date_time'])


fl_dask = fl_dask.sort_values('Date_time')
mt_dask = mt_dask.sort_values('Date_time')


## Se realiza un merge entre los dataframes de vuelos y metars por el Date_time, cogiendo el metar justamente anterior

%%time
fl_dask = fl_dask.sort_values('Date_time')
mtr_dask = mt_dask.sort_values('Date_time')

# Realiza el merge asof para encontrar el parte meteorológico más cercano en el tiempo
fl_dask = dd.merge_asof(fl_dask, mt_dask, on='Date_time', direction='backward')


pd.set_option('display.max_columns', None)

fl_dask.head()

fl_dask = fl_dask.sort_values('Date_time', ascending=False)
fl_dask.head()

fl_dask = fl_dask.rename(columns={'day': 'Date'})

fl_dask.info()

forma_df = fl_dask.shape[0].compute(), fl_dask.shape[1]
print("Forma del DataFrame:", forma_df)

fl_dask.to_csv("../data/flights/flights_with_metars.csv", index = False)

fl_dask.to_parquet("../data/flights/flights_with_metars.gz", compression='gzip', write_index=False)

fl_dask.to_json("../data/flights/flights_with_metars.json")
