==> Jupyter_Notebooks/5.1_database_creation.ipynb <==
# En este Jupyter Notebook se va a crear y cargar la base de datos en MySQL mediante SQLalchemy

# !pip install sqlalchemy

# !pip install mysql-connector-python

from sqlalchemy import create_engine, Column, Integer, String, Date, ForeignKey
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import relationship
from datetime import datetime
import pandas as pd
from sqlalchemy.orm import sessionmaker
with open('pass/pass.txt', 'r') as file:
    password = file.read().strip()

# Creación de la base de datos en MySQL

Base = declarative_base()

# Crear la base de datos 'mad_flights2' si no existe
database_creation_sql = "CREATE DATABASE IF NOT EXISTS mad_flights2;"
engine_creation = create_engine(f'mysql+mysqlconnector://root:{password}@localhost:3306/')
with engine_creation.connect() as connection:
    connection.execute(database_creation_sql)

# Actualizar la URL de la base de datos con el nombre correcto
database_url = f'mysql+mysqlconnector://root:{password}@localhost:3306/mad_flights2'

# Crear la conexión al motor de la base de datos
engine = create_engine(database_url, echo=True)

class Metar(Base):
    __tablename__ = 'metars'

    Metar_id = Column(Integer, primary_key=True, autoincrement=True)
    Date_time = Column(Date, default=None, nullable=True)
    Day = Column(Date, default=None, nullable=True)
    Hour = Column(String(45), default=None, nullable=True)
    Condition = Column(String(45), default=None, nullable=True)
    Temperature = Column(Integer, default=None, nullable=True)
    Wind = Column(Integer, default=None, nullable=True)
    Gusts = Column(Integer, default=None, nullable=True)
    Relative_hum = Column(Integer, default=None, nullable=True)
    Pressure = Column(Integer, default=None, nullable=True)

class Flight(Base):
    __tablename__ = 'flights'

    departure_date_time = Column(Date, default=None, nullable=True)
    flight_id = Column(Integer, primary_key=True, autoincrement=True)
    metar_id = Column(Integer, ForeignKey('metars.Metar_id'))
    cod_flight_IATA = Column(String(45), default=None, nullable=True)
    cod_flight_ICAO = Column(String(45), default=None, nullable=True)
    day = Column(Date, default=None, nullable=True)
    week_day = Column(String(45), default=None, nullable=True)
    status = Column(String(45), default=None, nullable=True)
    airliner = Column(String(45), default=None, nullable=True)
    cod_airliner_IATA = Column(String(45), default=None, nullable=True)
    cod_airliner_ICAO = Column(String(45), default=None, nullable=True)
    Scheduled_dep = Column(String(45), default=None, nullable=True)
    depart_time = Column(String(45), default=None, nullable=True)
    dep_situation = Column(String(45), default=None, nullable=True)
    dep_mins_of_delay = Column(String(45), default=None, nullable=True)
    city = Column(String(45), default=None, nullable=True)
    cod_airport_IATA = Column(String(45), default=None, nullable=True)
    cod_airport_ICAO = Column(String(45), default=None, nullable=True)
    arrival = Column(String(45), default=None, nullable=True)
    arr_situation = Column(String(45), default=None, nullable=True)
    arr_mins_of_delay = Column(String(45), default=None, nullable=True)
    duration = Column(String(45), default=None, nullable=True)
    resta = Column(Integer, default=None, nullable=True)

    metar = relationship('Metar', back_populates='flights')

Metar.flights = relationship('Flight', back_populates='metar')

# Crear las tablas en la base de datos
Base.metadata.create_all(engine)

# Carga de los datos en la tabla metars

# Rutas a tus archivos CSV
metars_csv_path = "../data/metars/metars_2017_2023.csv"


# Configuración de la base de datos
database_url = f'mysql+mysqlconnector://root:{password}@localhost:3306/mad_flights2'

# Crear el motor de la base de datos
engine = create_engine(database_url, echo=True)

# Cargar datos en la tabla 'metars'
metars_data = pd.read_csv(metars_csv_path)
metars_data.to_sql('metars', con=engine, if_exists='append', index=False)



# Carga de los datos en la tabla flights

# Ruta al archivo CSV
flights_csv_path = "../data/flights/flights_with_metarid.csv"

# Configuración de la base de datos
database_url = f'mysql+mysqlconnector://root:{password}@localhost:3306/mad_flights2'

# Crear el motor de la base de datos
engine = create_engine(database_url, connect_args={"connect_timeout": 300, "autocommit": True})

# Cargar datos en la tabla 'flights' en lotes más pequeños
chunksize = 100  # Puedes ajustar este valor según tus necesidades

# Crear un lector de CSV para iterar sobre lotes
csv_reader = pd.read_csv(flights_csv_path, chunksize=chunksize)

# Iterar sobre lotes e insertar en la base de datos
for i, chunk in enumerate(csv_reader):
    # Agregar lógica de procesamiento o limpieza si es necesario
    chunk.to_sql('flights', con=engine, if_exists='append', index=False)
